backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
firebase-admin==6.4.0
plaid-python==9.0.0
google-generativeai==0.3.2
langchain==0.0.350
langchain-google-genai==0.0.6
pydantic==2.5.0
httpx==0.25.2
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
